make v02 promoter collection
